A Priori

The apriori function implements the A Priori algorithm for association rule mining first proposed by Rakesh Agrawal and Srikant Ramakrishnan in 1994. This algorithm identifies frequent itemsets in a dataset and generates association rules based on specified support thresholds.

RuleMiner.apriori — Method

apriori(txns::Transactions, min_support::Union{Int,Float64}, max_length::Int)::DataFrame

Identify association rules in a transactional dataset txns, with minimum support, min_support, and maximum rule length, max_length.

When an Int value is supplied to min_support, apriori will use absolute support (count) of transactions as minimum support.

When a Float value is supplied, it will use relative support (percentage).

Parameters

txns::Transactions: A Transactions type object that contains the encoded transaction dataset as a sparse CSC matrix along with row and column name keys

min_support::Union{Int,Float64}: The minimum support threshold for the rules. This can be specified as either:

An Int represents the absolute support (count) of transactions.

A Float64 represents the relative support (percentage) of transactions.

max_length::Int: An integer value specifying the maximum length of the rules generated by the function

Output

The function returns a DataFrame containing the discovered association rules. Each row in the DataFrame represents a rule with the following columns:

LHS: The left-hand side of the rule (antecedent)

RHS: The right-hand side of the rule (consequent)

Support: Relative support of the rule

Confidence: Confidence of the rule

Coverage: Coverage (RHS support) of the rule

Lift: Lift of the assocation rule

N: Absolute support of association rule

Length: Length of the association rule

Algorithm Overview

The function starts by identifying base nodes (single-item rules) that meet the minimum support criteria.

It then iteratively generates longer rules up to the specified max_length.

For each rule length, it generates candidate rules by combining shorter rules with additional candidate items.

These candidate rules are then evaluated on whether they meet the minimum support citeria.

Various metrics (support, confidence, coverage, lift) are calculated for each rule which meets the min_support criterion.

Performance Considerations

This function uses Julia's multithreading capabilities to parallelize the generation and evaluation of candidate nodes, which can significantly improve performance on multi-core systems with Julia's native multithreading enabled.

Usage Example

# Load transactions
txns = load_transactions("transactions.txt", ' ')

# Find frequent itemsets with minimum support of 5% and maximum length of 3
result = apriori(txns, 0.05,3)