LevelWise

The levelwise function implements the levelwise algorithm for mining closed itemsets proposed by Pasquier et al in 1999. This algorithm generates all subsets of the closed itemsets, derives their supports, and then returns the results. This particular implementation is designed to take a result datafrom from the various closed itemset mining algorithms in this package as its input.

RuleMiner.levelwise — Method

levelwise(df::DataFrame, min_n::Int)::DataFrame

Recover frequent itemsets from a dataframe df of closed itemsets with minimum absolute support min_n

Parameters

df::DataFrame: A DataFrame DataFrame object with four columns:

Itemset: Vector of item names in the frequent itemset

Support: Relative support of the itemset

N: Absolute support count of the itemset

Length: Number of items in the itemset

min_support::Int: The minimum support threshold for the rules. This algorithm only takes absolute (integer) support

Output

A DataFrame object with three columns:

Itemset: Vector of item names in the frequent itemset

N: Absolute support count of the itemset

Length: Number of items in the itemset

Algorithm Overview

The function starts by creating candidates from all subcombinations of the closed itemsets

It computes their supports by finding the smallest closed itemset that is a superset of the candiate

The algorithm loops thorugh, building larger combinations until there are no combinations left

Usage Example

# Load transactions
txns = load_transactions("transactions.txt", ' ')

# Find frequent itemsets with minimum support of 5%
closed_sets = fpclose(txns, 0.05)

# Recover all frequent itemsets from the result
levelwise(closed_sets,0.05)