Transactions Objects

Transactions

struct Transactions
    matrix::SparseMatrixCSC{Bool,Int64}
    colkeys::Dict{Int,String} 
    linekeys::Dict{Int,String} 
end

The Transactions struct consists of three fields:

matrix: Sparse matrix showing the locations of the items (columns) in the transactions(rows)

colkeys: Dictionary mapping column indexes to their original values in the source

linekeys: Dictionary mapping line indexes to their original values in the source (or generated index number)

load_transactions

Note: this reads basket-format data, not tabular data!

RuleMiner.load_transactions — Method

load_transactions(file::String, delimiter::Char; id_col::Bool= false, skiplines::Int= 0, nlines::Int= 0)::Transactions

Read transaction data from a file where each line is a list of items separated by a given delimiter

If the first item of each list is a transaction identifier, set id_col to true

Specify the number header lines to skip with skiplines

Specify a specific number of lines to read with nlines

Reads transaction data from a file basket-format file, where each line is a list of items, and returns a Transactions object.

Parameters:

file: Path to the input file

delimiter: Character used to separate items in each transaction

id_col: Set to true if the first item in each line is a transaction identifier (default: false)

skiplines: Number of header lines to skip (optional)

nlines:Number of lines to read (optional)

Returns:

Transactions: A Transactions object with a sparse matrix representing the treansactions (rows) and items (columns) as well as dictionary keys for the names of the rows and columns

transactions

RuleMiner.transactions — Method

transactions(df::DataFrame; indexcol::Union{Symbol,Nothing}= nothing)::Transactions

Converts a one-hot encoded DataFrame object into a Transactions object

Designate a column as an index column with indexcol

Converts a one-hot encoded DataFrame into a Transactions object.

Parameters:

df: Input DataFrame

indexcol: Column to use as the index (optional)

txns_to_df

RuleMiner.txns_to_df — Method

txns_to_df(txns::Transactions; id_col::Bool= false)::DataFrame

Convert a Transactions object to a DataFrame. Specify whether the Transactions linekey field should be included as an Index column with indexcol

Parameters:

txns: The Transactions object to be converted.

id_col: If true, an additional 'Index' column is added to the DataFrame containing the values from the linekeys dictionary.

Returns

DataFrame: A one-hot encoded DataFrame representation of the Transactions object.

Usage Examples

# Load transactions from a file
txns = load_transactions("transactions.txt", ' ', id_col=true, skiplines=1)

# Convert a DataFrame to Transactions
df = DataFrame(Index = [1,2,3], A=[1,0,1], B=[0,1,1], C=[1,1,0])

txns = transactions(df, indexcol=:Index)